Reorg codegen under a common subdir
