--- src/js/mod.rs
@@ -1,2 +1,0 @@
-pub mod builder;
-pub mod printer;
